requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyascii==0.3.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
asyncio-throttle==1.0.2
asyncpg==0.27.0
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
boto3==1.26.160
botocore==1.29.160
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
defusedxml==0.7.1
Django==4.2.2
django-bakery==0.13.2
django-debug-toolbar==4.1.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
fs==2.4.16
gevent==22.10.2
gevent-eventemitter==2.1
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-resources==5.12.0
jmespath==1.0.1
kiwisolver==1.4.4
l18n==2021.3
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pkg_resources==0.0.0
plotly==5.14.1
protobuf==3.20.3
psycopg2-binary==2.9.6
pycryptodomex==3.18.0
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
steam==1.4.4
telepath==0.3.1
tenacity==8.2.2
typing_extensions==4.6.3
urllib3==1.26.16
vdf==3.4
wagtail==5.0.2
wagtail-bakery==0.8.0
webencodings==0.5.1
Willow==1.5
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
